refactor: deprecate and make fs.cachedChecks noop
