+++ packages/vite/src/node/server/middlewares/htmlFallback.ts
@@ -1,8 +1,7 @@
 import path from 'node:path'
+import fs from 'node:fs'
 import type { Connect } from 'dep-types/connect'
 import { createDebugger } from '../../utils'
-import type { FsUtils } from '../../fsUtils'
-import { commonFsUtils } from '../../fsUtils'
 import { cleanUrl } from '../../../shared/utils'
 
 const debug = createDebugger('vite:html-fallback')
@@ -10,7 +9,6 @@ const debug = createDebugger('vite:html-fallback')
 export function htmlFallbackMiddleware(
   root: string,
   spaFallback: boolean,
-  fsUtils: FsUtils = commonFsUtils,
 ): Connect.NextHandleFunction {
   // Keep the named function. The name is visible in debug logs via `DEBUG=connect:dispatcher ...`
   return function viteHtmlFallbackMiddleware(req, _res, next) {
@@ -37,7 +35,7 @@ export function htmlFallbackMiddleware(
     // so we need to check if the file exists
     if (pathname.endsWith('.html')) {
       const filePath = path.join(root, pathname)
-      if (fsUtils.existsSync(filePath)) {
+      if (fs.existsSync(filePath)) {
         debug?.(`Rewriting ${req.method} ${req.url} to ${url}`)
         req.url = url
         return next()
@@ -46,7 +44,7 @@ export function htmlFallbackMiddleware(
     // trailing slash should check for fallback index.html
     else if (pathname[pathname.length - 1] === '/') {
       const filePath = path.join(root, pathname, 'index.html')
-      if (fsUtils.existsSync(filePath)) {
+      if (fs.existsSync(filePath)) {
         const newUrl = url + 'index.html'
         debug?.(`Rewriting ${req.method} ${req.url} to ${newUrl}`)
         req.url = newUrl
@@ -56,7 +54,7 @@ export function htmlFallbackMiddleware(
     // non-trailing slash should check for fallback .html
     else {
       const filePath = path.join(root, pathname + '.html')
-      if (fsUtils.existsSync(filePath)) {
+      if (fs.existsSync(filePath)) {
         const newUrl = url + '.html'
         debug?.(`Rewriting ${req.method} ${req.url} to ${newUrl}`)
         req.url = newUrl
